Move crd resource syncer tests into separate files
