+++ tests/fiaas_deploy_daemon/crd/test_crd_watcher.py
@@ -97,28 +97,8 @@ class TestWatcher(object):
         return mock.create_autospec(spec=Lifecycle, spec_set=True, instance=True)
 
     @pytest.fixture
-    def apiextensions_v1_crd_watcher(self, spec_factory, deploy_queue, watcher, lifecycle):
-        crd_watcher = CrdWatcher(spec_factory, deploy_queue, Configuration(["--use-apiextensionsv1-crd"]), lifecycle,
-                                 FakeCrdResourcesSyncer)
-        crd_watcher._watcher = watcher
-        return crd_watcher
-
-    @pytest.fixture
     def crd_watcher(self, spec_factory, deploy_queue, watcher, lifecycle):
         crd_watcher = CrdWatcher(spec_factory, deploy_queue, Configuration([]), lifecycle, FakeCrdResourcesSyncer)
-        crd_watcher._watcher = watcher
-        return crd_watcher
-
-    @pytest.fixture
-    def apiextensions_v1_crd_watcher_with_syncer(self, spec_factory, deploy_queue, watcher, lifecycle):
-        crd_watcher = CrdWatcher(spec_factory, deploy_queue, Configuration(["--use-apiextensionsv1-crd"]), lifecycle,
-                                 CrdResourcesSyncerApiextensionsV1)
-        crd_watcher._watcher = watcher
-        return crd_watcher
-
-    @pytest.fixture
-    def apiextensions_v1beta1_crd_watcher_with_syncer(self, spec_factory, deploy_queue, watcher, lifecycle):
-        crd_watcher = CrdWatcher(spec_factory, deploy_queue, Configuration([]), lifecycle, CrdResourcesSyncerApiextensionsV1Beta1)
         crd_watcher._watcher = watcher
         return crd_watcher
 
@@ -128,209 +108,10 @@ class TestWatcher(object):
             m.side_effect = NotFound
             yield m
 
-    def test_creates_apiextensions_v1_crd_if_not_exists_when_watching_it(self, get, post, apiextensions_v1_crd_watcher_with_syncer,
-                                                                         watcher):
-        get.side_effect = NotFound("Something")
-        watcher.watch.side_effect = NotFound("Something")
-
-        object_with_unknown_fields = {"type": "object", "x-kubernetes-preserve-unknown-fields": True}
-        expected_application = {
-            'metadata': {
-                'namespace': 'default',
-                'name': 'applications.fiaas.schibsted.io',
-                'ownerReferences': [],
-                'finalizers': [],
-            },
-            'spec': {
-                'group': 'fiaas.schibsted.io',
-                'names': {
-                    'shortNames': ['app', 'fa'],
-                    'kind': 'Application',
-                    'plural': 'applications',
-                    'categories': []
-                },
-                'preserveUnknownFields': False,
-                'scope': 'Namespaced',
-                'conversion': {
-                    'strategy': 'None'
-                },
-                'versions': [{
-                    'additionalPrinterColumns': [],
-                    'name': 'v1',
-                    'served': True,
-                    'storage': True,
-                    'schema': {
-                        'openAPIV3Schema': {
-                            'type': 'object',
-                            'properties': {
-                                'spec': {
-                                    'type': 'object',
-                                    'properties': {
-                                        "application": {
-                                            "type": "string",
-                                        },
-                                        "image": {
-                                            "type": "string",
-                                        },
-                                        "config": object_with_unknown_fields,
-                                        "additional_labels": {
-                                            "type": "object",
-                                            "properties": {
-                                                "global": object_with_unknown_fields,
-                                                "deployment": object_with_unknown_fields,
-                                                "horizontal_pod_autoscaler": object_with_unknown_fields,
-                                                "ingress": object_with_unknown_fields,
-                                                "service": object_with_unknown_fields,
-                                                "service_account": object_with_unknown_fields,
-                                                "pod": object_with_unknown_fields,
-                                                "status": object_with_unknown_fields,
-                                            }
-                                        },
-                                        "additional_annotations": {
-                                            "type": "object",
-                                            "properties": {
-                                                "global": object_with_unknown_fields,
-                                                "deployment": object_with_unknown_fields,
-                                                "horizontal_pod_autoscaler": object_with_unknown_fields,
-                                                "ingress": object_with_unknown_fields,
-                                                "service": object_with_unknown_fields,
-                                                "service_account": object_with_unknown_fields,
-                                                "pod": object_with_unknown_fields,
-                                                "status": object_with_unknown_fields,
-                                            }
-                                        }
-                                    }
-                                }
-                            },
-                            'oneOf': [],
-                            'allOf': [],
-                            'required': [],
-                            'x-kubernetes-list-map-keys': [],
-                            'anyOf': []
-                        }
-                    }
-                }]
-            }
-        }
-        expected_status = {
-            'metadata': {
-                'namespace': 'default',
-                'name': 'application-statuses.fiaas.schibsted.io',
-                'ownerReferences': [],
-                'finalizers': [],
-            },
-            'spec': {
-                'group': 'fiaas.schibsted.io',
-                'names': {
-                    'shortNames': ['status', 'appstatus', 'fs'],
-                    'kind': 'ApplicationStatus',
-                    'plural': 'application-statuses',
-                    'categories': []
-                },
-                'preserveUnknownFields': False,
-                'scope': 'Namespaced',
-                'conversion': {
-                    'strategy': 'None'
-                },
-                'versions': [{
-                    'additionalPrinterColumns': [],
-                    'name': 'v1',
-                    'served': True,
-                    'storage': True,
-                    'schema': {
-                        'openAPIV3Schema': {
-                            'type': 'object',
-                            'properties': {
-                                "result": {
-                                    "type": "string"
-                                },
-                                "logs": {
-                                    "type": "array",
-                                    "items": {
-                                        "type": "string"
-                                    }
-                                }
-                            },
-                            'oneOf': [],
-                            'allOf': [],
-                            'required': [],
-                            'x-kubernetes-list-map-keys': [],
-                            'anyOf': []
-                        }
-                    }
-                }]
-            }
-        }
-
-        def make_response(data):
-            mock_response = mock.create_autospec(Response)
-            mock_response.json.return_value = data
-            return mock_response
-
-        post.side_effect = [make_response(expected_application), make_response(expected_status)]
-
-        apiextensions_v1_crd_watcher_with_syncer._watch(None)
-
-        calls = [
-            mock.call("/apis/apiextensions.k8s.io/v1/customresourcedefinitions/", expected_application),
-            mock.call("/apis/apiextensions.k8s.io/v1/customresourcedefinitions/", expected_status)
-        ]
-        assert post.call_args_list == calls
-
-    def test_creates_apiextensions_v1beta1_crd_if_not_exists_when_watching_it(self, get, post, watcher,
-                                                                              apiextensions_v1beta1_crd_watcher_with_syncer):
-        get.side_effect = NotFound("Something")
-        watcher.watch.side_effect = NotFound("Something")
-
-        expected_application = {
-            'metadata': {
-                'namespace': 'default',
-                'name': 'applications.fiaas.schibsted.io',
-                'ownerReferences': [],
-                'finalizers': [],
-            },
-            'spec': {
-                'version': 'v1',
-                'group': 'fiaas.schibsted.io',
-                'names': {
-                    'shortNames': ['app', 'fa'],
-                    'kind': 'Application',
-                    'plural': 'applications'
-                }
-            }
-        }
-        expected_status = {
-            'metadata': {
-                'namespace': 'default',
-                'name': 'application-statuses.fiaas.schibsted.io',
-                'ownerReferences': [],
-                'finalizers': [],
-            },
-            'spec': {
-                'version': 'v1',
-                'group': 'fiaas.schibsted.io',
-                'names': {
-                    'shortNames': ['status', 'appstatus', 'fs'],
-                    'kind': 'ApplicationStatus',
-                    'plural': 'application-statuses'
-                }
-            }
-        }
-
-        def make_response(data):
-            mock_response = mock.create_autospec(Response)
-            mock_response.json.return_value = data
-            return mock_response
-
-        post.side_effect = [make_response(expected_application), make_response(expected_status)]
-
-        apiextensions_v1beta1_crd_watcher_with_syncer._watch(None)
-
-        calls = [
-            mock.call("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/", expected_application),
-            mock.call("/apis/apiextensions.k8s.io/v1beta1/customresourcedefinitions/", expected_status)
-        ]
-        assert post.call_args_list == calls
+    def test_updates_crd_resources_when_watching_it(self, crd_watcher):
+        with mock.patch.object(FakeCrdResourcesSyncer, "update_crd_resources", return_value=None) as m:
+            crd_watcher._watch(None)
+            m.assert_called_once()
 
     def test_is_able_to_watch_custom_resource_definition(self, crd_watcher, deploy_queue, watcher):
         watcher.watch.return_value = [WatchEvent(ADD_EVENT, FiaasApplication)]
